+++ codelab_101/minsang/lib/favorites_page.dart
@@ -1,0 +1,29 @@
+import 'package:codelab_101/main.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+class FavoritesPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    var appState = context.watch<MyAppState>();
+
+    if (appState.favorites.isEmpty) {
+      return Center(
+        child: Text('No favorites yet.'),
+      );
+    }
+
+    return ListView(
+      children: [
+        Padding(
+          padding: const EdgeInsets.all(20),
+          child: Text('You have '
+              '${appState.favorites.length} favorites:'),
+        ),
+        for (var pair in appState.favorites)
+          ListTile(
+              leading: Icon(Icons.favorite), title: Text(pair.asLowerCase)),
+      ],
+    );
+  }
+}
